test: move unit test into st directory
